cloud_recog/requirements.txt
flask==3.0.2
numpy==1.26.4
opencv-python-headless==4.9.0.80
insightface==0.7.3
python-dotenv==1.0.1
uwsgi==2.0.23
onnxruntime-gpu==1.16.0
PyJWT==2.6.0
